requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
audioop-lts==0.2.1
audioread==3.0.1
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.4
billiard==4.2.1
binaryornot==0.4.4
black==25.1.0
bleach==6.2.0
boto3==1.38.38
botocore==1.38.38
build==1.4.0
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
chromadb==1.4.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cookiecutter==2.6.0
cron-descriptor==1.4.5
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
Django==5.2.2
django-browser-reload==1.18.0
django-celery-beat==2.8.1
django-filter==25.1
django-storages==1.14.6
django-tailwind==4.0.1
django-timezone-field==7.1
django-unfold==0.59.0
django_celery_results==2.6.0
djangorestframework==3.16.1
docx==0.2.4
dotenv==0.9.9
drf-spectacular==0.28.0
durationpy==0.10
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
filelock==3.18.0
flatbuffers==25.12.19
fonttools==4.58.4
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2026.1.0
future==1.0.0
google-auth==2.47.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface_hub==1.3.1
humanfriendly==10.0
identify==2.6.12
idna==3.10
importlib_metadata==8.7.1
importlib_resources==6.5.2
inflection==0.5.1
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
kombu==5.5.4
kubernetes==34.1.0
langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.74
langchain-openai==0.3.14
langchain-text-splitters==0.3.9
langgraph==0.2.69
langgraph-checkpoint==2.1.1
langgraph-checkpoint-postgres==2.0.23
langgraph-sdk==0.1.74
langsmith==0.3.45
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.44.0
loguru==0.7.3
lxml==6.0.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
mutagen==1.47.0
mypy_extensions==1.1.0
narwhals==1.44.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.9.1
notebook==7.4.4
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.23.2
openai==1.101.0
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.2
ormsgpack==1.10.0
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
pooch==1.8.2
posthog==5.4.0
pre_commit==4.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.33.4
psutil==7.0.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
redis==6.2.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.0
scikit-learn==1.7.0
scipy==1.16.0
Send2Trash==1.8.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
sql_metadata==2.17.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.11.0
tinycss2==1.4.0
tokenizers==0.22.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.21.1
typer-slim==0.21.1
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
virtualenv==20.31.2
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
whitenoise==6.9.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
faiss-cpu==1.9.0.post1
